requirements.txt
Pillow
pika
requests
